refactor(triggers): relocate dispatch modules out of the legacy outbox tree
dispatchError -> server/triggers/ (shared by all sender modules);
emailHourlyCap -> app-layer/triggers/dispatch/emailCaps.ts;
triggerActionDispatch + graphAlertActionDispatch -> app-layer/triggers/dispatch/.
Imports rewritten at every consumer; tests move alongside their modules.
Prepares the legacy outbox directory for deletion (ADR-052).
