+++ langwatch/src/server/event-sourcing/queues/groupQueue/groupQueue.ts
@@ -26,7 +26,7 @@ import {
   type ProjectStorageDestination,
   redactStorageUrisInText,
 } from "../../../stored-objects/project-storage-destination";
-import { isDispatchError } from "../../outbox/dispatchError";
+import { isDispatchError } from "../../../triggers/dispatchError";
 import type {
   DeduplicationConfig,
   EventSourcedQueueDefinition,
