+++ langwatch/src/server/triggers/__tests__/sendWebhook.dispatch.unit.test.ts
@@ -1,5 +1,5 @@
 import { afterEach, describe, expect, it, vi } from "vitest";
-import { DispatchError } from "~/server/event-sourcing/outbox/dispatchError";
+import { DispatchError } from "~/server/triggers/dispatchError";
 
 // Stub the SSRF-fenced transport and the rate limiter so these tests exercise
 // sendWebhook's ORCHESTRATION (event-id header, dispatch cap, Retry-After
