+++ langwatch/src/server/triggers/__tests__/slackWebApi.unit.test.ts
@@ -1,5 +1,5 @@
 import { afterEach, describe, expect, it, vi } from "vitest";
-import { DispatchError } from "~/server/event-sourcing/outbox/dispatchError";
+import { DispatchError } from "~/server/triggers/dispatchError";
 import { sendHttpDestination } from "../httpDestination";
 import { listSlackChannels, postSlackChatMessage } from "../slackWebApi";
